Move shared lib code to internal/
